requirements.txt
requests
pyyaml
gtts
moviepy